GEOMESA-144,GEOMESA-140
GEOMESA-140 Warn users about missing temporal component of index
This commit adds a warning when a user neglects to properly set SF_PROPERTY_START_TIME in the UserData. Such a mistake is non-fatal, but does impact performance.

GEOMESA-144 Set temporal index variable by default if not selected by user
This commit will set SF_PROPERTY_START_TIME for the user if a time attribute is present in the SimpleFeatureType, and emit a notification to this effect.