WIP GEOMESA-265, 264 Create a tools wrapper shell script and Make output for `list` command line tool prettier
This is just a basic wrapper shell script for Geomesa Tools,
but it comes with some important features. The first is that it can
help with configuring environment variables through the `geomesa configure`
command. This will prompt the user to set the required variables if
they aren't already set. It will also put `geomesa` on the path and set up
GEOMESA_HOME based on the root directory of geomesa. Now, one can run commands
just by typing `geomesa ...` instead of running commands through the JAR file.

Additionally, the output of `geomesa list` was improved to show more details and
appear prettier and easier to read.