GEOMESA-641 Add raster query profiling and dumpQueries geomesa-tools command
Raster queries are now profiled on the getRasters, Accumulo scanning, and raster mosaicing steps
of returning raster chunks to a client. These are then written to an Accumulo table and can be dumped from that table
with the "dumpQueries" geomesa-tools command. This should aid in debugging and configuring environmental setup options.