Forward-port proper spark partitioning behavior
This adds an rddPartPerExec function that allows the specification
the number of parts per executor. This relies on some fairly rickety
functionality, so I've taken the liberty of marking it pre-emptively
with the deprecated decorator. This capability should eventually
be firmed up and added to the core "rdd" function to better
map to the normative pattern for spark APIs. And, you know,
move away from hilariously tightly dependent code.