Bug fixes to split grouping size calculation
The previous approach was a hold over from when we wanted to push the
number of splits around so that it was based on the geomesa shards.

This is now fixed to match the intent: produce K partitions per
active executor, unless there are too few ranges. We do not current
split large ranges into small ones, which is an eventual goal.