GEOMESA-1402 Kafka DS secondary indexing via CQEngine
This PR creates a new module called ``geomesa-memory`` which provides
the basis of a new implementation of ``LiveFeatureCache`` (the
interface of which is split out as a trait). The original implementation
of this cache is now renamed ``LiveFeatureCacheGuava`` and is still
used by default; the new cache may be enabled via a KDS parameter.

The geomesa-memory cache is based on CQEngine, which provides the
ability to fulfill queries via a set of configurable attribute indexes,
falling back to iteration when indexes are not available. The
``BucketIndex`` used in ``LiveFeatureCacheGuava`` has been added as a
new CQEngine geo-index; other geographic operators are (for now) handled
by iteration. CQEngine is also configured to logically deduplicate
results from queries with overlapping OR clauses.

Jim Hughes <jnh5y@ccri.com>, Matthew Zimmerman <matt.zimmerman@ccri.com>
with some help from John Armstrong <john.armstrong@ccri.com>

Signed-off-by: Matthew Zimmerman <matt.zimmerman@ccri.com>