GEOMESA-1463 Fix case-sensitive filename collisions
by removing many of the PNG files present which are not actually linked
to in the corresponding tutorial RST, including the ones that are
identical on case-insensitive filesystems.

Also converted the progression.ogv movie to the animated GIF
progression.gif, which can be rendered out of the box by most browsers.
This file is larger (11MB vs 4MB) than the old one, but the increase in
size is mostly offset by the other unused files removed.

Signed-off-by: Matthew Zimmerman <matt.zimmerman@ccri.com>