GEOMESA-1658 Bundle Vegas for Jupyter+Toree.
This moves the current geomesa-jupyter to geomesa-jupyter-leaflet,
and adds geomesa-jupyter-vegas, which builds a shaded JAR with
all dependencies needed to run Vegas in Jupyter+Toree. This can
be added, along with the GeoMesa Spark JARs, at Toree configuration
time. Also adds brief instructions to the Spark chapter of the docs
on use.

Requires the ``vegas`` profile to build:

    mvn clean install -Pvegas -pl geomesa-jupyter/geomesa-jupyter-vegas

Signed-off-by: Matthew Zimmerman <matt.zimmerman@ccri.com>