GEOMESA-2447 Refactor the CartesianProductIterator to remove costly modulus calls
Reworked the implementation so that it no longer uses the integer
division, but does a more direct iteration over array indexes.
The JIRA ticket contains a chart that shows, for the few test
cases run, the new implementation takes approximately half as
long to generate and exhaust the iterators over the product.

Signed-off-by: Chris Eichelberger <cne1x@ccri.com>