GEOMESA-2478 Update git-commit-id-plugin to latest version.
Upgrade git-commit-id-plugin from 2.2.1 to the latest version,
2.2.5, in the root POM.

The previous versions of the plugin were hanging for me
when trying to build.  Upgrading fixed that issue for me.
This might have to do with the following bug:

https://github.com/git-commit-id/maven-git-commit-id-plugin/issues/336

Note: the version in the root POM was last updated from 2.1.13 to
2.2.1 in May 2017.

In 3 other POMs, remove git-commit-id-plugin <version> which was set
to 2.1.13 (perhaps originally copied from the root POM):

 * geomesa-accumulo-tools/pom.xml,
   geomesa-lambda-tools/pom.xml: version 2.1.13 was selected
   in the initial commits of these POMs in Sep 2016 and July 2017.
   The latter may have been copy-pasted from the former.

 * geomesa-fs-tools/pom.xml: version 2.1.13 was selected in the
   git-commit-id-plugin configuration when it was added to this
   build in Sep 2017... perhaps copy-paste from one of the above
   two POMs.

Signed-off-by: Carsten Clark <github-cdc@ccri.com>