GEOMESA-2478 Upgrade scala-maven-plugin from 3.2.1 to 3.4.4 in the root POM.
This version of the plugin requires at least Maven 3.3.9.
Update README.md accordingly.  Delete the incomplete building
instructions in CONTRIBUTING.md and refer to README.md instead.

Since version 3.3.1, the scala-maven-plugin's way of invoking
Zinc changed so that Zinc may look in the wrong path for the
"javac" command.  Since version 3.3.2, this problem can be
defeated by setting the JAVA_HOME environment variable:

https://github.com/davidB/scala-maven-plugin/pull/227

Add a note about this to the README because it caused me trouble.

No significant diffs in the console output of a full build.

Signed-off-by: Carsten Clark <github-cdc@ccri.com>