GEOMESA-2478 Update Maven plugins to current versions (#2117)
* GEOMESA-2478 Update git-commit-id-plugin to latest version.

Upgrade git-commit-id-plugin from 2.2.1 to the latest version,
2.2.5, in the root POM.

The previous versions of the plugin were hanging for me
when trying to build.  Upgrading fixed that issue for me.
This might have to do with the following bug:

https://github.com/git-commit-id/maven-git-commit-id-plugin/issues/336

Note: the version in the root POM was last updated from 2.1.13 to
2.2.1 in May 2017.

In 3 other POMs, remove git-commit-id-plugin <version> which was set
to 2.1.13 (perhaps originally copied from the root POM):

 * geomesa-accumulo-tools/pom.xml,
   geomesa-lambda-tools/pom.xml: version 2.1.13 was selected
   in the initial commits of these POMs in Sep 2016 and July 2017.
   The latter may have been copy-pasted from the former.

 * geomesa-fs-tools/pom.xml: version 2.1.13 was selected in the
   git-commit-id-plugin configuration when it was added to this
   build in Sep 2017... perhaps copy-paste from one of the above
   two POMs.

Signed-off-by: Carsten Clark <github-cdc@ccri.com>

* GEOMESA-2478 Update all core Maven plugins to their latest versions.

Update all maven-*-plugin entries in the root POM's pluginManagement
to the latest version for each plugin.  Latest version info obtained
from https://maven.apache.org/plugins/ which turned out to match
what was available in Maven Central at
http://repo.maven.apache.org/maven2/org/apache/maven/plugins/

Add a comment to each maven-*-plugin configuration saying when
the version was released, to make it easier to spot potentially
out-of-date plugins.

Upgrading the maven-compiler-plugin required turning off its
new 'useIncrementalCompilation' option in order to keep it from
recompiling Java files that the scala-maven-plugin already compiled.

I see no significant differences in the output of "build/mvn install"
before and after this change (other than the plugin version numbers).

Signed-off-by: Carsten Clark <github-cdc@ccri.com>

* GEOMESA-2478 Add pluginManagement entries for 3 missing core Maven plugins.

Again no significant diffs in the console output of a full build.

Signed-off-by: Carsten Clark <github-cdc@ccri.com>

* GEOMESA-2478 Bump license-maven-plugin from 2.10 to 3.10 in root POM.

3.0 was released 2016-08-20 and is the latest as of 2018-11-30.

Again no significant diffs in the console output of a full build.

Signed-off-by: Carsten Clark <github-cdc@ccri.com>

* GEOMESA-2478 Upgrade scala-maven-plugin from 3.2.1 to 3.4.4 in the root POM.

This version of the plugin requires at least Maven 3.3.9.
Update README.md accordingly.  Delete the incomplete building
instructions in CONTRIBUTING.md and refer to README.md instead.

Since version 3.3.1, the scala-maven-plugin's way of invoking
Zinc changed so that Zinc may look in the wrong path for the
"javac" command.  Since version 3.3.2, this problem can be
defeated by setting the JAVA_HOME environment variable:

https://github.com/davidB/scala-maven-plugin/pull/227

Add a note about this to the README because it caused me trouble.

No significant diffs in the console output of a full build.

Signed-off-by: Carsten Clark <github-cdc@ccri.com>