GEOMESA-2726 add the application of Google S2 algorithm in point data (#2370)
The implementation of S2 is completely implemented by Google's algorithm; s3 I use epoch+S2+Time to encode, so although the 3D is not uniformly coded, the performance of the actual test is very good. We also analyzed the reasons, mainly in the 64-bit encoding. If the accuracy of each dimension is 3, the accuracy of each dimension will be 21, and the number of results of the scan will be more. More than the final result, through the above S3 encoding, we have tested this scan+mutiRowFilter to extract data from HBase with a higher accuracy than Z3, of course not absolute, most of the scenarios we tested.
In addition, very excited, this is the first time I submitted a PR to the open source community. If there is something wrong, please correct me.

Signed-off-by: jhkcool <jhkcool@qq.com>