Arrow: introduce an axis order parameter
Geomesa currently defaults to always reading and writing Arrow geometries in Lat/Lon order, with the new axis order parameter allow the user to request the encoding to be explicitly specified per request.

Axis Order parameter supported values:
- LatLon translates to: Latitude = Y Axis = North, Longitude = X Axis = East
- LonLat translates to: Longitude = X Axis = East, Latitude = Y Axis = North

The request parameter is optional and if not specified axis order defaults to the existing behavior which is LatLon.